chore(sonar): clear 8 mechanical code_smells (S1905 + S6213)
S1905 x3 — remove unnecessary long/int → float casts in ChurnFeatureExtractor.

S6213 x5 — rename test variables named record (contextual keyword in
Java 16+) to rec.

+++ src/main/java/org/iris/ml/ChurnFeatureExtractor.java
@@ -88,14 +88,14 @@ public class ChurnFeatureExtractor {
         int domainClass = classifyEmailDomain(customer.getEmail());
 
         return new float[] {
-                (float) daysSinceLast,
+                daysSinceLast,  // long → float implicit widening (Sonar S1905)
                 (float) rev30,
                 (float) rev90,
                 (float) rev365,
                 (float) frequency,
                 (float) diversity,
-                (float) domainClass,
-                (float) lifetimeDays,
+                domainClass,    // int → float implicit widening (Sonar S1905)
+                lifetimeDays,   // long → float implicit widening (Sonar S1905)
         };
     }
